Move generated files to their own dir
