+++ packages/apps/docs/.gitignore
@@ -1,7 +1,5 @@
 # generated resources
-src/data/searchIndex.json
-src/data/menu.mjs
-src/data/mostPopular.json
+src/_generated
 src/pages/docs/kadena/kadenajs
 src/pages/docs/kadena/client
 src/pages/docs/pact/cli
@@ -18,7 +16,5 @@ src/pages/docs/chainweb/stream-client
 src/pages/docs/chainweb/node-client
 src/pages/docs/chainweb/js-bindings
 
-
 next-env.d.ts
 .env.local
-
