+++ packages/apps/docs/package.json
@@ -11,7 +11,7 @@
     "build:brokenLinks": "node src/scripts/detectBrokenLinks.mjs",
     "build:checkForHeaders": "node src/scripts/checkForHeaders.mjs && prettier ./src/pages/docs --write",
     "build:createSpecJson": "node ./src/scripts/createSpec.mjs && prettier ./src/specs/**/*.json --write",
-    "build:createtree": "node ./src/scripts/getdocstree.mjs && prettier ./src/data/menu.mjs --write",
+    "build:createtree": "node ./src/scripts/getdocstree.mjs && prettier ./src/_generated/menu.mjs --write",
     "build:importReadMe": "node ./src/scripts/importReadme.mjs && prettier ./src/pages/docs --write",
     "build:next": "pnpm run build:scripts && next build",
     "build:scripts": "pnpm run build:importReadMe && pnpm run build:createtree && pnpm run build:createSpecJson && pnpm run build:brokenLinks && pnpm run build:checkForHeaders",
